requirements.txt
discord.py==1.7.3
odmantic
dnspython
dpytools
DiscordUtils
discord-pretty-help
python-dotenv
requests